coderpal/pyproject.toml
[tool.poetry]
name = "coderpal"
version = "0.1.0"
description = ""
authors = ["Rohan Almeida <rohan.almeida@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
ipython = "^8.11.0"
python-dotenv = "^1.0.0"
black = "^23.1.0"
openai = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
